Version 1.0.172-83
* Fixed bug with Polygon Triangulation where the last triangle wasn't added to the list of triangles
* Added a `copy_obstacles_from()` function to copy the obstacles of another object
* Added object tagging which allows the user to access a GraphicsObject through tags (can be strings, ints, etc.)
* Objects now store their instances in sets to make the code faster
* The Animated Image class now calls the `_move()` function of the Image objects and not the `move()` function
* Performance Upgrades! tested 2400 frames on test game with cProfile:

      FPS increased from ~650 FPS to ~3040 FPS

      BEFORE:--------------------------

      Trial 1: 4.025 secs, 596.273 FPS
      Trial 2: 3.959 secs, 606.214 FPS
      Trial 3: 3.663 secs, 655.201 FPS
      Trial 4: 3.013 secs, 796.548 FPS
      Trial 5: 3.864 secs, 621.118 FPS

      Average: 3.705 secs, 647.808 FPS

      AFTER: --------------------------

      Trial 1: 0.871 secs, 2755.454 FPS
      Trial 2: 0.686 secs, 3498.542 FPS
      Trial 3: 0.749 secs, 3204.272 FPS
      Trial 4: 0.928 secs, 2586.207 FPS
      Trial 5: 0.712 secs, 3370.787 FPS

      Average: 0.789 secs, 3041.054 FPS

* The _BBox objects' `is_clicked()` functions run about 2841.3% (~285x) faster faster!
* Added `get_height()` and `get_width()` functions to the Polygon class
* The Polygon `get_height()` and `get_width()` functions no longer recalculate the height & width

* Reorganized the if statement arguments to reduce processing time in the `_on_mouse_motion()` function
* Transform class `screen()` function now returns a float object
* The GraphicsObject doesn't iterate through all the objects now, only the ones that are required to improve performance

+++ goopylib/colours.py
@@ -48,9 +48,9 @@ class Colour:
             green = self.green + other.green
             blue = self.blue + other.blue
         elif isinstance(other, int) or isinstance(other, float):
-            red = int(self.red + other)
-            green = int(self.green + other)
-            blue = int(self.blue + other)
+            red = (self.red + other) // 1
+            green = (self.green + other) // 1
+            blue = (self.blue + other) // 1
         else:
             raise TypeError(f"unsupported operand type(s) for +: 'Colour' and {type(other)}'. Must be a Colour or int or float")
 
@@ -62,9 +62,9 @@ class Colour:
             green = self.green * other.green
             blue = self.blue * other.blue
         elif isinstance(other, int) or isinstance(other, float):
-            red = int(self.red * other)
-            green = int(self.green * other)
-            blue = int(self.blue * other)
+            red = (self.red * other) // 1
+            green = (self.green * other) // 1
+            blue = (self.blue * other) // 1
         else:
             raise TypeError(f"unsupported operand type(s) for *: 'Colour' and {type(other)}'. "
                             f"Must be a Colour or int or float")
@@ -77,9 +77,9 @@ class Colour:
             green = self.green // other.green
             blue = self.blue // other.blue
         elif isinstance(other, int) or isinstance(other, float):
-            red = int(self.red // other)
-            green = int(self.green // other)
-            blue = int(self.blue // other)
+            red = (self.red // other) // 1
+            green = (self.green // other) // 1
+            blue = (self.blue // other) // 1
         else:
             raise TypeError(f"unsupported operand type(s) for //: 'Colour' and {type(other)}'. Must be a Colour or int or float")
 
@@ -94,9 +94,9 @@ class Colour:
             green = self.green % other.green
             blue = self.blue % other.blue
         elif isinstance(other, int) or isinstance(other, float):
-            red = int(self.red % other)
-            green = int(self.green % other)
-            blue = int(self.blue % other)
+            red = (self.red % other) // 1
+            green = (self.green % other) // 1
+            blue = (self.blue % other) // 1
         else:
             raise TypeError(f"unsupported operand type(s) for %: 'Colour' and {type(other)}'. Must be a Colour or int or float")
 
@@ -108,9 +108,9 @@ class Colour:
             green = pow(self.green ** power.green, modulo)
             blue = pow(self.blue ** power.blue, modulo)
         elif isinstance(power, int) or isinstance(power, float):
-            red = int(pow(self.red ** power, modulo))
-            green = int(pow(self.green ** power, modulo))
-            blue = int(pow(self.blue ** power, modulo))
+            red = (pow(self.red ** power, modulo)) // 1
+            green = (pow(self.green ** power, modulo)) // 1
+            blue = (pow(self.blue ** power, modulo)) // 1
         else:
             raise TypeError(f"unsupported operand type(s) for **: 'Colour' and {type(power)}'. Must be a Colour or int or float")
